+++ post/admin.py
@@ -3,4 +3,5 @@ from django.contrib import admin
 from jmbo.admin import ModelBaseAdmin
 from post.models import Post
 
+
 admin.site.register(Post, ModelBaseAdmin)