+++ post/models.py
@@ -1,28 +1,34 @@
+import markdown
 from bs4 import BeautifulSoup
 
+from django.db import models
 from django.core.urlresolvers import reverse
-from ckeditor.fields import RichTextField
+
+from simplemde.fields import SimpleMDEField
 
 from jmbo.models import ModelBase
 
 
 class Post(ModelBase):
-    autosave_fields = ('content',)
+    autosave_fields = ("markdown",)
 
-    content = RichTextField(
-        blank=True,
-        null=True,
-    )
+    markdown = SimpleMDEField(null=True, blank=True)
 
     class Meta:
-        verbose_name = 'Post'
-        verbose_name_plural = 'Posts'
+        verbose_name = "Post"
+        verbose_name_plural = "Posts"
+
+    @property
+    def content(self):
+        if not self.markdown:
+            return ""
+        return markdown.markdown(self.markdown)
 
     @property
     def content_pages(self):
-        marker = '--m-a-r-k-er--'
-        soup = BeautifulSoup(self.content)
-        elems = soup.find_all('div', attrs={'style': 'page-break-after: always;'})
+        marker = "--m-a-r-k-er--"
+        soup = BeautifulSoup(self.content, "html.parser")
+        elems = soup.find_all("hr")
         for elem in elems:
             elem.replace_with(marker)
         return soup.renderContents().split(marker)