+++ post/tests/urls.py
@@ -1,14 +1,22 @@
-from django.conf.urls import patterns, include
+from django.conf.urls import include, url
 from django.contrib import admin
+
+from rest_framework import routers
+from rest_framework_extras import discover
+
+from post import api as post_api
 
 
 admin.autodiscover()
 
-urlpatterns = patterns(
-    '',
-    (r'^jmbo/', include('jmbo.urls')),
-    (r'^comments/', include('django.contrib.comments.urls')),
-    (r'^ckeditor/', include('ckeditor.urls')),
-    (r'^post/', include('post.urls')),
-    (r'^admin/', include(admin.site.urls)),
-)
+router = routers.SimpleRouter()
+post_api.register(router)
+discover(router)
+
+urlpatterns = [
+    url(r"^admin/", include(admin.site.urls)),
+    url(r'^api/(?P<version>(v1))/', include(router.urls)),
+    url(r"^jmbo/", include("jmbo.urls", namespace="jmbo")),
+    url(r"^post/", include("post.urls", namespace="jmbo")),
+    url(r"^comments/", include("django_comments.urls")),
+]