+++ CHANGELOG.rst
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+3.0.0
+-----
+#. Jmbo 3 and Django 1.9 compatibility.
+#. Store content as markdown, not HTML anymore.
+
 2.0.0
 -----
 #. Simplify API to use primary keys.